### 2019-09-06 #1715 @kyleniemeyer
@arfon I don't know anything about R, but can cover this if needed
