### 2019-09-06 #1715 @arfon
> @arfon I don't know anything about R, but can cover this if needed

Yeah, sorry about this. I'm struggling with finding editors to cover submissions right now sorry.
