### 2019-09-24 #1715 @kyleniemeyer
Hi @corybrunson @vc1492a @camillescott @szhorvat, are any of you be interested and available to review this submission for JOSS?
