### 2019-09-24 #1715 @corybrunson
@kyleniemeyer sure, i have some background both in graph theory and in R, and i'll be interested to see how the algorithm compares. I should be able to do a first full pass within 5 days.
