### 2019-09-25 #1715 @corybrunson
@kyleniemeyer i'm sorry if i'm just confused, but i'm wondering why i was given access to `openjournals/joss-reviews`. I reflexively accepted it, thinking it was the review repo, and then started getting updates on reviews and pre-reviews from all of JOSS. I've unwatched the repo but i want to check whether i was supposed to have been granted access. Thanks!

P.S. I'm still down to review this submission!
