### 2019-09-25 #1715 @kyleniemeyer
@corybrunson yes, you are given temporary access so that you can check off items in the review issue, when it is started—which will happen when I have a second reviewer assigned. Unwatching the repo is definitely what we recommend.
