### 2019-09-26 #1766 @whedon
[REVIEW]: yenpathy: An R Package to Quickly Find K Shortest Paths Through a Weighted Graph
**Submitting author:** @toph-allen (<a href="http://orcid.org/0000-0003-4580-091X">Toph Allen</a>)
**Repository:** <a href="https://github.com/ecohealthalliance/yenpathy" target ="_blank">https://github.com/ecohealthalliance/yenpathy</a>
**Version:** 1.0.0
