### 2019-09-27 #1766 @szhorvat
@kyleniemeyer  Since it appears that this package was born partly because R/igraph could not satisfy a need, I should disclose that I am an active contributor to the C core of igraph (not the R interface).
