### 2019-09-27 #1766 @szhorvat
I would have liked to see this functionality be contributed directly into the igraph C core (this is not meant as a criticism though!), and we _might_ add similar functionality to igraph after getting inspired by `yenpathy`, so that it could be used by igraph interfaces in other languages as well.
