### 2019-09-27 #1766 @szhorvat
@toph-allen

Minor language comments on the paper:

In "Summary", "Langauge" -> "Language"

"Csardi" -> "Csárdi"

In the Performance section, you use quotation marks on "simple" in _"simple" paths_. This is in fact standard graph theory terminology.
