### 2019-09-27 ecohealthalliance/yenpathy#5 @szhorvat
What kinds of graphs does it work with?
This is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1766).

An important requirement for any package dealing with graphs is to be clear about what types of graphs it works with, and ensure that it behaves reasonably (return a non-surprising result or issue an error) for any type of graph that it is given.

My usual checklist:

### Undirected vs directed

It appears that this package considers everything to be directed, but this is not documented anywhere. It is especially disconcerting to pass the function an undirected igraph object and see it handled as directed (with arbitrary direction assignments).

This must be corrected.

### Graphs with multi-edges

Consider this graph:

![image](https://user-images.githubusercontent.com/1212871/65754491-754a2d80-e111-11e9-90fb-ec0fa6d6c851.png)

How many paths from 1 to 3? One or two?

Again, this must be at least documented.

### Graph with self-loops

`k_shortest_paths(g, 100, 100)` returns the result `100 100` regardless of whether the vertex `100` is even present in the graph. It does not take into account whether the vertex has a self-loop.

Again, in the very least both behaviours (self-loop handling and handling of non-resent vertices) should be documented.

I understand that a dataframe is unable to encode isolated vertices, which might be grounds for considering any vertex to be present. But this is surprising behaviour that should be clearly documented with examples.

----

Generally, if the function takes a certain datatype as input (e.g. an igraph graph) then it must reasonably handle all possible forms this datatype can take (directed/undirected, multigraphs, self-loops, empty graphs, weighted/unweighted, etc.)
