### 2019-09-27 #1766 @szhorvat
Added comments on the types of graphs that the function handles: https://github.com/ecohealthalliance/yenpathy/issues/5
