### 2019-09-27 #1766 @toph-allen
@szhorvat Thanks for the text edits and the detailed comments on the issue! I'll take a look in greater depth in the next day or so.
