### 2019-10-01 #1766 @corybrunson
Hi @toph-allen, thanks for building the package! I was unfamiliar with Yen's algorithm and am enjoying seeing it play out.

The software paper is rich with examples but begins with a minimal justificatory paragraph. This is appropriate to the light weight of the package. I have a minor suggestion: to change "a method to find the _k_ shortest paths through a network" to "a method to find the shortest several paths through a network, up to any number _k_". (To be fair, my own awkward reading at first was as a specialist in a different area, where important parameterized concepts are often called "_k_-things", and i was wondering what a "_k_-shortest path" was.)
