### 2019-10-01 ecohealthalliance/yenpathy#7 @corybrunson
handling of integer versus character node IDs
One of the consistent behaviors of the igraph R package is the correct handling of node IDs whether given as characters or as numerics, and i expect that R users familiar with igraph will expect this behavior from yenpathy as well. While it seems appropriate not to extend it to data frame inputs, the following behavior should, i think, either be amended or clarified with a warning:

``` r
library(yenpathy)
data(karate, package = "igraphdata")
k_shortest_paths(karate, from = "Mr Hi", to = "John A")
#> Loading required namespace: igraph
#> [[1]]
#> [1] "Mr Hi"    "Actor 20" "John A"
k_shortest_paths(karate, from = 1, to = 34)
#> [[1]]
#> [1] NA NA
```

Moreover, in the case of a data frame input, the behavior is asymmetric, and neither output is informative:

``` r
library(yenpathy)
g <- data.frame(
  start = I(letters[c(1, 2, 4)]),
  end = I(letters[c(2, 3, 5)])
)
k_shortest_paths(g, from = 1, to = 3)
#> [[1]]
#> [1] NA NA
g <- data.frame(
  start = c(1, 2, 4),
  end = c(2, 3, 5)
)
k_shortest_paths(g, from = "a", to = "c")
#> Error in .k_shortest_paths_Cpp(graph_df, from, to, k, vertex_num, verbose): Not compatible with requested type: [type=character; target=integer].
```

Finally, `from` and `to` seem to be coerced to characters when the node ID columns are character encodings of integers, rather than prompting an error or warning:

``` r
library(yenpathy)
g <- data.frame(
  start = I(as.character(c(3, 4, 1))),
  end = I(as.character(c(4, 5, 2)))
)
k_shortest_paths(g, from = 1, to = 3)
#> list()
```

Is this intentional?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1766)._
