### 2019-10-01 #1766 @corybrunson
@kyleniemeyer apologies; i meant to move my previous comment to an issue on the repo, to avoid clutter here in the review thread, and i'll do so now.
