### 2019-10-01 #1766 @kyleniemeyer
@corybrunson no worries, it's fine either way
