### 2019-10-02 #1766 @toph-allen
@corybrunson I think your suggested change for the first paragraph is definitely clearer, and will push that change up. I'm going to discuss your other comments on the issues in the repo. Thanks!
