### 2019-10-02 #1766 @corybrunson
@toph-allen cool, glad it suits you. I may not respond to updates this workweek but i'll be back to the review this weekend.
