### 2019-10-02 ecohealthalliance/yenpathy#5 @toph-allen
@szhorvat Thanks for the detailed comment — I'm fairly new to working with graphs, and your explanation makes it all very clear.

The C++ code we use as the core of the package assumes that all graphs are directed, and the work we're using this for deals only with directed graphs, so those other cases you describe weren't on our radar.

I'll update the package's code and documentation to deal with those cases. Where it's possible to handle them seamlessly, I'll do that. If there are instances where handling a case would require a rewrite of the C++ code that's beyond what I'm able to do at this point, I'll explicitly check for that case and handle it as gracefully as possible.

For undirected graphs: The C++ code takes a data frame as its input, and assumes that each edge is directed. Would representing an undirected graph using a data frame of mutual edges cause any issues? If so, it's probably better to explicitly state that the current version of the package handles directed graphs only, even though as you state this is not ideal.

I'll reply with any other questions as they occur to me. Thanks again!
