### 2019-10-02 ecohealthalliance/yenpathy#6 @toph-allen
Yeah, I had opted not to handle factors because I thought that — because of the trickiness of factors' dual integer and character representation — it was less likely that they'd be used to represent a graph.

But to your point, I can imagine a scenario where someone would read in a graph from a CSV file with `read.csv()`, which would use factors by default if there were characters in the node names, and be unable to use yenpathy's functions.

I'll add some bookkeeping code to convert factors on the way in and out of the function.
