### 2019-10-02 ecohealthalliance/yenpathy#8 @toph-allen
Agreed on both points — I'll amend them to use fixed graphs and to run at a higher value of `k`.
