### 2019-10-02 ecohealthalliance/yenpathy#5 @szhorvat
> representing an undirected graph using a data frame of mutual edges

I think that's the correct solution here.
