### 2019-10-02 8f2c7d5
Minor changes to text of paper.Rmd.
