+++ inst/paper/paper.bib
@@ -13,7 +13,7 @@
 }
 @article{igraph,
   title = {The igraph software package for complex network research},
-  author = {Gabor Csardi and Tamas Nepusz},
+  author = {Gabor Csárdi and Tamas Nepusz},
   journal = {InterJournal},
   volume = {Complex Systems},
   pages = {1695},
