+++ inst/paper/paper.Rmd
@@ -29,8 +29,9 @@ journal: JOSS
 
 Finding the shortest paths through a network is a task important to many
 problems, from transportation routing to social network analysis. **yenpathy**
-provides a method to find the _k_ shortest paths through a network using Yen's
-Algorithm [@Yen_1971], previously not available in the R Langauge [@R].
+provides a method to find several shortest paths through a network, up to any
+number _k_, using Yen's Algorithm [@Yen_1971], previously not available in the
+R Language [@R].
 
 **yenpathy** provides the function `k_shortest_paths()`, which returns *k*
 shortest paths between two nodes in a graph, ordered from shortest to longest,
@@ -123,7 +124,7 @@ sapply(itineraries, function(.x) sum(.x[["Distance"]]))
 ### Performance
 
 A number of packages, such as **igraph** and **dodgr**,  provide functions to 
-find the shortest path between two nodes on a network, or *all* "simple" (non-looping)
+find the shortest path between two nodes on a network, or *all* simple (non-looping)
 paths between nodes. For the *single shortest* path, these packages can be faster than
 **yenpathy**:
 
