### 2019-10-06 #1766 @szhorvat
@kyleniemeyer

What is the procedure to follow when a reviewer requests changes? Should I write up the review as-is for now (to complete it on time), or should I wait until the authors address the concerns?

Since the review is open, I assume that the review will keep going continuously until all concerns are addressed (even if this may take a long time). Let me know if this is not how it works.
