### 2019-10-06 #1766 @kyleniemeyer
@szhorvat yes, the review process is meant to be iterative and keep going until concerns are addressed—but this isn't meant to work as slowly as a traditional process where the authors get all reviewer comments at once, prepare a rebuttal, and repeat, but instead more like a conversation.

Feel free to leave comments here or in the software repo as you work through your review. Hope that answers your question.
