### 2019-10-07 ecohealthalliance/yenpathy#9 @szhorvat
Result may be incorrect when the input is a multigraph
This is related to #5.

Consider the following dataframe, which encodes a multigraph:

```
df<-data.frame(source=c(1,1,2,1), sink=c(2,2,3,3), weight = c(1,4,1,3))
```

The shortest path from 1 to 3 is 1->2->3 where we need to take the 1->2 edge with weight 1 (not the one with weight 4). Thus the total path length is 1 (for 1->2) + 1 (for 2->3) = 2.

However, the function (incorrectly) returns 1->3

```
> k_shortest_paths(df,1,3)
[[1]]
[1] 1 3
```

It appears that the first 1->2 edge (with a small weight) may be internally discarded.

Note that users of the package may pass in a multigraph completely inadvertently.
