### 2019-10-07 #1766 @szhorvat
@corybrunson wrote:

> (To be fair, my own awkward reading at first was as a specialist in a different area, where important parameterized concepts are often called "k-things", and i was wondering what a "k-shortest path" was.)

I know exactly what you mean. An easy fix (to avoid this misunderstanding) would be to use `n` instead of `k`.
