### 2019-10-07 #1766 @szhorvat
The first stage of my review is complete. I am now waiting for the comments to be addressed.

### Summary

This R package provides a function to find the `k` shortest paths between two given vertices of a graph. It relies on [an existing C++ library](https://github.com/yan-qi/k-shortest-paths-cpp-version). The main contribution is exposing the functionality to R, and making it convenient to use from R.

### Ease of use and integration with R

I do not use R often, and I do not feel qualified to judge how well integrated the function is. However, @corybrunson has already written multiple comments on this.

Installation was easy on macOS (the only platform where I tested it), but it did require compiling from source. It may be useful to mention what is required for a successful installation (e.g. Xcode being installed on macOS). I am also wondering how easy it would be to get it working on Windows, where a compiler isn't typically available.

### Functionality

A typical issue with software that works with graphs is that the authors did not consider all types of graphs which could be given as input. I opened two issues regarding this:

 - https://github.com/ecohealthalliance/yenpathy/issues/5
 - https://github.com/ecohealthalliance/yenpathy/issues/9

These should be addressed in some way before publishing the package.

### Documentation

The paper contains more examples than the documentation. However, users will read the docs, not the paper. I recommend expanding the docs with these examples.

It would be useful to illustrate common tasks such as:

 - getting the path in terms of edges (with their weights)
 - getting the total length for each path

On a related note:

Have you considered returning a data structure which does not simply contain the paths as lists of vertices, but also lists of edges (e.g. edge indices = dataframe row indices), as well as the total length of each path?  R/igraph functions typically return all the information that was already computed.

### Minor comments

 - It would be useful to mention the time complexity of this implementation (in terms of the number of vertices and edges).

 - In the docs,

    > `edge_penalty` | A constant to be added to THE WEIGHT OF each edge

 - igraph should be spelt as "igraph", not "iGraph"

 - When solving combinatorial problems that may have a very large number of solutions, it is often useful to have some sort of "streaming" or "iterator" interface. I.e., once we found `k` paths and examined them, we may decide that we want the `k+1`th path. A "streaming interface" would make it possible to do this without having to re-run the entire function.  (Of course, this is a minor point, and I do not expect that you will implement this.)
