### 2019-10-07 #1766 @toph-allen
@szhorvat Thanks for the review! It's helpful to receive comments from someone with more experience in this field than I do.

We'll make the minor copy changes you suggest.

Responding to your other comments:
- In R, the general convention is for the top-level README.md to contain a high-level overview, with more elaborate examples going in the package vignettes. This package contains one vignette which is at parity with the paper. We could add another vignette with the examples you suggest. We can also refer users to the examples from the README.md.
- We'll investigate the bug reports you filed.
- We considered adding additional functionality to return the length of the paths. During our internal development, we tried to keep the functionality as light-weight and simple as possible for the sake of getting it done fast, but returning the length of the paths seems like it would be a useful feature.
- I can definitely see how a streaming interface would be useful, but it would probably require significant changes to the C++ code, which we have till now avoided rewriting significantly.
