### 2019-10-07 #1766 @szhorvat
@toph-allen

> This package contains one vignette which is at parity with the paper.

Thanks for the explanation! Could you let me know how to access it? I was using the R Studio interface to browse the documentation and I do not see it there (for other packages I do). I also tried

```
library(yenpathy)
vignette(package='yenpathy')
```

but it says `no vignettes found`.

I am not experienced with R.
