### 2019-10-07 #1766 @corybrunson
Hi @szhorvat and @toph-allen, just in case this isn't obvious (it wasn't to me until i figured it out), R package vignettes may only become available when a package is installed with a specific flag for it:

```r
# install and build vignettes
remotes::install_github("ecohealthalliance/yenpathy", build_vignettes = TRUE)
# list available vignettes
vignette(package = "yenpathy")
# open specific vignette
vignette("using-yenpathy")
```
