### 2019-10-07 #1766 @szhorvat
Thanks @corybrunson, it works this way. Indeed it would be nice to add it to the installation instructions.
