### 2019-10-10 874ce11
Set up pkgdown and gh-pages.
