+++ .travis.yml
@@ -10,3 +10,9 @@ r_packages:
 
 after_success:
   - Rscript -e "library(covr); codecov()"
+
+before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+  deploy:
+    provider: script
+    script: Rscript -e 'pkgdown::deploy_site_github()'
+    skip_cleanup: true
