+++ .travis.yml
@@ -8,11 +8,11 @@ warnings_are_errors: false
 r_packages:
   - covr
 
+before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github()'
+  skip_cleanup: true
+
 after_success:
   - Rscript -e "library(covr); codecov()"
-
-before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
-  deploy:
-    provider: script
-    script: Rscript -e 'pkgdown::deploy_site_github()'
-    skip_cleanup: true
