+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: '2.2'
+pkgdown: 1.4.1
+pkgdown_sha: ~
+articles:
+  using-yenpathy: using-yenpathy.html
+
