### 2019-10-10 #1766 @toph-allen
I added the vignette-building argument and viewing instructions to README.Rmd.

I also used [`pkgdown`](https://github.com/r-lib/pkgdown) to create a [GitHub Pages site](https://ecohealthalliance.github.io/yenpathy/) which includes the README, R documentation, and vignettes, which you can view here, and added a link to this site to the package's README.md.
