### 2019-10-10 878dd9c
Added vignette instructions to README.Rmd.
