### 2019-10-21 #1766 @szhorvat
@toph-allen @kyleniemeyer I am just giving advance notice that I may not be able to respond or comment here during most of November (travelling in a place where GitHub is blocked).
