### 2019-11-01 #1766 @toph-allen
Sorry for my going awol on this thread — I had a few deadlines come up which have made working on this tough. I’m going to try to begin addressing some of the issues in the next week.

> On Oct 21, 2019, at 8:36 AM, Szabolcs Horvát <notifications@github.com> wrote:
>
> @toph-allen <https://github.com/toph-allen> @kyleniemeyer <https://github.com/kyleniemeyer> I am just giving advance notice that I may not be able to respond or comment here during most of November (travelling in a place where GitHub is blocked).
>
