### 2019-11-23 #1766 @toph-allen
@kthyng Thanks for checking in. I've made some minor progress, but some unexpected work deadlines have meant that making changes to Yenpathy have had to take a back burner. I'll do my best to have some commits pushed this week.
