### 2020-06-08 fe19461
Update README.md
add specification of "directed" graph
