+++ pkgdown.yml
@@ -1,6 +1,7 @@
 pandoc: '2.2'
-pkgdown: 1.4.1
+pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   using-yenpathy: using-yenpathy.html
+last_built: 2020-06-08T16:58Z
 
