### 2020-09-15 #1766 @toph-allen
Sorry, this email slipped down in my inbox. We decided to withdraw Yenpathy from submission to JOSS — it seems like it's a little under the scope of what qualifies for the journal, and the excellent feedback from the reviewers suggest that to bring it up to snuff, we'd have to do a fair amount of reworking of how it approaches the problem.

The project started life as a kinda hacky way to accomplish something we needed — that is, find an arbitrary number of shortest paths quickly through a large directed graph. When I wrote it, I was able to make a lot of assumptions about the structure of the graph, and didn't need to be as rigorous handling different kinds of edge cases.

Adding analysis of graphs passed in to the package's functions is outside of the scope of our original intent. Furthermore, to properly debug the code, I'd need to work in more depth with the C++ library we're wrapping, and that's beyond my abilities. I think it makes more sense for the package to stand "as-is".

Thanks again to JOSS folks and the reviewers — even if the package isn't getting submitted to the journal, your feedback and even just preparing it for submission has improved the quality of the package and its documentation.
